requirements.txt
PySide6>=6.5
Pillow>=10.0.0
